+++ src/libs/ImageWithCaption.jsx
@@ -5,6 +5,7 @@ import PhotoIcon from '@material-symbols/svg-400/rounded/photo.svg';
  * @param {object} props
  * @param {string} props.src
  * @param {string} props.alt
+ * @param {string} [props.caption]
  * @param {string} [props.className]
  * @param {'add'} [props.usage]
  * @param {import('react').ReactNode} [props.children]
@@ -14,6 +15,7 @@ export default function ImageWithCaption({
   className,
   src,
   alt,
+  caption,
   usage,
   children,
   Icon = getIconByUsage(usage),
@@ -30,9 +32,11 @@ export default function ImageWithCaption({
       ) : (
         <Icon className="flex-1 fill-gray-400" />
       )}
-      <figcaption className="absolute right-2 bottom-0 text-right font-mono [text-shadow:_-1px_-1px_2px_white,_-1px_1px_2px_white,_1px_1px_2px_white,_1px_-1px_2px_white]">
-        {alt}
-      </figcaption>
+      {caption && (
+        <figcaption className="absolute right-2 bottom-0 text-right font-mono [text-shadow:_-1px_-1px_2px_white,_-1px_1px_2px_white,_1px_1px_2px_white,_1px_-1px_2px_white]">
+          {caption}
+        </figcaption>
+      )}
       {children}
     </figure>
   );